Refactor main method call
